revert to single binary, runtime checks only
