+++ cmd/zitadel.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"bytes"
 	"context"
+	"crypto/fips140"
 	_ "embed"
 	"errors"
 	"io"
@@ -27,6 +28,9 @@ var (
 
 	//go:embed defaults.yaml
 	defaultConfig []byte
+
+	//go:embed defaults_fips.yaml
+	defaultFipsConfig []byte
 )
 
 func New(out io.Writer, in io.Reader, args []string, server chan<- *start.Server) *cobra.Command {
@@ -77,6 +81,17 @@ func loadDefaultConfigInto(v *viper.Viper) error {
 		return err
 	}
 	return mergeFipsDefaultConfig(v)
+}
+
+func mergeFipsDefaultConfig(v *viper.Viper) error {
+	if !fips140.Enabled() {
+		return nil
+	}
+	return applyFipsDefaultOverlay(v)
+}
+
+func applyFipsDefaultOverlay(v *viper.Viper) error {
+	return v.MergeConfig(bytes.NewBuffer(defaultFipsConfig))
 }
 
 func initConfig() {
