--- cmd/zitadel_nofips.go
@@ -1,7 +1,0 @@
-//go:build !fips
-
-package cmd
-
-import "github.com/spf13/viper"
-
-func mergeFipsDefaultConfig(*viper.Viper) error { return nil }
