+++ nx.json
@@ -66,18 +66,6 @@
         "default"
       ],
       "cache": true
-    },
-    "test-unit-fips": {
-      "inputs": [
-        "default"
-      ],
-      "cache": true
-    },
-    "test-integration-fips": {
-      "inputs": [
-        "default"
-      ],
-      "cache": true
     }
   }
 }
